Break apart Data.js [pre- AS-577]
