+++ src/components/IGVBrowser.js
@@ -3,10 +3,10 @@ import { Fragment, useRef, useState } from 'react'
 import { div, h } from 'react-hyperscript-helpers'
 import { requesterPaysWrapper } from 'src/components/bucket-utils'
 import { Link } from 'src/components/common'
+import { getUserProjectForWorkspace, parseGsUri } from 'src/components/data/data-utils'
 import { centeredSpinner, icon } from 'src/components/icons'
 import { Ajax, saToken } from 'src/libs/ajax'
 import colors from 'src/libs/colors'
-import { getUserProjectForWorkspace, parseGsUri } from 'src/libs/data-utils'
 import { reportError } from 'src/libs/error'
 import { knownBucketRequesterPaysStatuses, requesterPaysProjectStore } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
