+++ src/components/WorkflowSelector.js
@@ -1,6 +1,7 @@
 import _ from 'lodash/fp'
 import { Fragment, useState } from 'react'
 import { div, h } from 'react-hyperscript-helpers'
+import { ModalToolButton } from 'src/components/data/data-utils'
 import { centeredSpinner } from 'src/components/icons'
 import { Ajax } from 'src/libs/ajax'
 import { withErrorReporting } from 'src/libs/error'
@@ -8,7 +9,6 @@ import * as Nav from 'src/libs/nav'
 import { workflowSelectionStore } from 'src/libs/state'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
-import { ModalToolButton } from 'src/pages/workspaces/workspace/Data'
 
 
 const WorkflowSelector = ({ workspace: { workspace: { namespace, name } }, selectedEntities }) => {
