+++ src/pages/workspaces/workspace/Data.js
@@ -1,46 +1,26 @@
-import * as clipboard from 'clipboard-polyfill/text'
-import FileSaver from 'file-saver'
 import filesize from 'filesize'
-import JSZip from 'jszip'
 import _ from 'lodash/fp'
-import * as qs from 'qs'
-import { Fragment, useEffect, useImperativeHandle, useRef, useState } from 'react'
-import { div, form, h, img, input } from 'react-hyperscript-helpers'
+import { Fragment, useEffect, useImperativeHandle, useState } from 'react'
+import { div, h } from 'react-hyperscript-helpers'
 import { AutoSizer } from 'react-virtualized'
 import * as breadcrumbs from 'src/components/breadcrumbs'
 import { requesterPaysWrapper, withRequesterPaysHandler } from 'src/components/bucket-utils'
 import Collapse from 'src/components/Collapse'
-import { ButtonPrimary, Clickable, Link, MenuButton, Select, spinnerOverlay } from 'src/components/common'
-import DataTable from 'src/components/DataTable'
+import { Clickable, Link, spinnerOverlay } from 'src/components/common'
+import { EntityUploader, ReferenceDataDeleter, ReferenceDataImporter, renderDataCell, saveScroll } from 'src/components/data/data-utils'
+import EntitiesContent from 'src/components/data/EntitiesContent'
+import LocalVariablesContent from 'src/components/data/LocalVariablesContent'
 import Dropzone from 'src/components/Dropzone'
-import ExportDataModal from 'src/components/ExportDataModal'
 import FloatingActionButton from 'src/components/FloatingActionButton'
-import { icon, spinner } from 'src/components/icons'
-import IGVBrowser from 'src/components/IGVBrowser'
-import IGVFileSelector from 'src/components/IGVFileSelector'
-import { DelayedSearchInput, TextInput } from 'src/components/input'
+import { icon } from 'src/components/icons'
+import { DelayedSearchInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
-import { withModalDrawer } from 'src/components/ModalDrawer'
-import { cohortNotebook, cohortRNotebook, NotebookCreator } from 'src/components/notebook-utils'
-import PopupTrigger from 'src/components/PopupTrigger'
 import { FlexTable, HeaderCell, SimpleTable, TextCell } from 'src/components/table'
-import TitleBar from 'src/components/TitleBar'
 import UriViewer from 'src/components/UriViewer'
-import WorkflowSelector from 'src/components/WorkflowSelector'
 import { SnapshotInfo } from 'src/components/workspace-utils'
-import datasets from 'src/data/datasets'
-import dataExplorerLogo from 'src/images/data-explorer-logo.svg'
-import igvLogo from 'src/images/igv-logo.png'
-import jupyterLogo from 'src/images/jupyter-logo.svg'
-import wdlLogo from 'src/images/wdl-logo.png'
 import { Ajax } from 'src/libs/ajax'
-import { getUser } from 'src/libs/auth'
 import colors from 'src/libs/colors'
-import { getConfig } from 'src/libs/config'
-import { EntityDeleter, EntityUploader, ReferenceDataDeleter, ReferenceDataImporter, renderDataCell } from 'src/libs/data-utils'
 import { withErrorReporting } from 'src/libs/error'
-import * as Nav from 'src/libs/nav'
-import { notify } from 'src/libs/notifications'
 import { pfbImportJobStore } from 'src/libs/state'
 import * as StateHistory from 'src/libs/state-history'
 import * as Style from 'src/libs/style'
@@ -67,33 +47,6 @@ const styles = {
   }
 }
 
-export const ModalToolButton = ({ icon, text, disabled, ...props }) => {
-  return h(Clickable, _.merge({
-    disabled,
-    style: {
-      color: disabled ? colors.secondary() : colors.accent(),
-      opacity: disabled ? 0.5 : undefined,
-      border: '1px solid transparent',
-      padding: '0 0.875rem', marginBottom: '0.5rem',
-      backgroundColor: 'white',
-      display: 'flex',
-      alignItems: 'center',
-      height: '3rem',
-      fontSize: 18,
-      userSelect: 'none'
-    },
-    hover: {
-      border: `1px solid ${colors.accent(0.8)}`,
-      boxShadow: Style.standardShadow
-    }
-  }, props), [
-    !!icon && div({ style: { display: 'flex', alignItems: 'center', width: 45, marginRight: '1rem' } }, [
-      img({ src: icon, style: { opacity: disabled ? 0.5 : undefined, maxWidth: 45, maxHeight: 40 } })
-    ]),
-    text
-  ])
-}
-
 const DataTypeButton = ({ selected, children, iconName = 'listAlt', iconSize = 14, buttonStyle, ...props }) => {
   return h(Clickable, {
     style: { ...Style.navList.item(selected), color: colors.accent(1.2), ...buttonStyle },
@@ -118,10 +71,6 @@ const DataImportPlaceholder = () => {
   ])
 }
 
-const saveScroll = _.throttle(100, (initialX, initialY) => {
-  StateHistory.update({ initialX, initialY })
-})
-
 const getReferenceData = _.flow(
   _.toPairs,
   _.filter(([key]) => key.startsWith('referenceData_')),
@@ -131,232 +80,6 @@ const getReferenceData = _.flow(
   }),
   _.groupBy('datum')
 )
-
-const LocalVariablesContent = ({ workspace, workspace: { workspace: { namespace, name } }, firstRender, refreshKey }) => {
-  const signal = Utils.useCancellation()
-
-  const [editIndex, setEditIndex] = useState()
-  const [deleteIndex, setDeleteIndex] = useState()
-  const [editKey, setEditKey] = useState()
-  const [editValue, setEditValue] = useState()
-  const [editType, setEditType] = useState()
-  const [textFilter, setTextFilter] = useState('')
-
-  const [busy, setBusy] = useState()
-  const [attributes, setAttributes] = useState()
-
-  const loadAttributes = _.flow(
-    withErrorReporting('Error loading workspace data'),
-    Utils.withBusyState(setBusy)
-  )(async () => {
-    const { workspace: { attributes } } = await Ajax(signal).Workspaces.workspace(namespace, name).details(['workspace.attributes'])
-    setAttributes(attributes)
-  })
-
-  useEffect(() => {
-    loadAttributes()
-  }, [refreshKey]) // eslint-disable-line react-hooks/exhaustive-deps
-
-  const stopEditing = () => {
-    setEditIndex()
-    setEditKey()
-    setEditValue()
-    setEditType()
-  }
-
-  const initialAttributes = _.flow(
-    _.toPairs,
-    _.remove(([key]) => /^description$|:|^referenceData_/.test(key))
-  )(attributes)
-
-  const creatingNewVariable = editIndex === initialAttributes.length
-  const amendedAttributes = _.flow(
-    _.filter(([key, value]) => Utils.textMatch(textFilter, `${key} ${value}`)),
-    _.sortBy(_.first),
-    arr => [...arr, ...(creatingNewVariable ? [['', '']] : [])]
-  )(initialAttributes)
-
-  const inputErrors = editIndex !== undefined && [
-    ...(_.keys(_.unset(amendedAttributes[editIndex][0], attributes)).includes(editKey) ? ['Key must be unique'] : []),
-    ...(!/^[\w-]*$/.test(editKey) ? ['Key can only contain letters, numbers, underscores, and dashes'] : []),
-    ...(editKey === 'description' ? ['Key cannot be \'description\''] : []),
-    ...(editKey.startsWith('referenceData_') ? ['Key cannot start with \'referenceData_\''] : []),
-    ...(!editKey ? ['Key is required'] : []),
-    ...(!editValue ? ['Value is required'] : []),
-    ...(editValue && editType === 'number' && Utils.cantBeNumber(editValue) ? ['Value is not a number'] : []),
-    ...(editValue && editType === 'number list' && _.some(Utils.cantBeNumber, editValue.split(',')) ?
-      ['Value is not a comma-separated list of numbers'] : [])
-  ]
-
-  const saveAttribute = _.flow(
-    withErrorReporting('Error saving change to workspace variables'),
-    Utils.withBusyState(setBusy)
-  )(async originalKey => {
-    const isList = editType.includes('list')
-    const newBaseType = isList ? editType.slice(0, -5) : editType
-
-    const parsedValue = isList ? _.map(Utils.convertValue(newBaseType), editValue.split(/,\s*/)) :
-      Utils.convertValue(newBaseType, editValue)
-
-    await Ajax().Workspaces.workspace(namespace, name).shallowMergeNewAttributes({ [editKey]: parsedValue })
-
-    if (editKey !== originalKey) {
-      await Ajax().Workspaces.workspace(namespace, name).deleteAttributes([originalKey])
-    }
-
-    await loadAttributes()
-    stopEditing()
-    setTextFilter('')
-  })
-
-  const upload = _.flow(
-    withErrorReporting('Error uploading file'),
-    Utils.withBusyState(setBusy)
-  )(async ([file]) => {
-    await Ajax().Workspaces.workspace(namespace, name).importAttributes(file)
-    await loadAttributes()
-  })
-
-  const download = _.flow(
-    withErrorReporting('Error downloading attributes'),
-    Utils.withBusyState(setBusy)
-  )(async () => {
-    const blob = await Ajax().Workspaces.workspace(namespace, name).exportAttributes()
-    FileSaver.saveAs(blob, `${name}-workspace-attributes.tsv`)
-  })
-
-  const { initialY } = firstRender ? StateHistory.get() : {}
-  return h(Dropzone, {
-    disabled: !!Utils.editWorkspaceError(workspace),
-    style: { flex: 1, display: 'flex', flexDirection: 'column' },
-    activeStyle: { backgroundColor: colors.accent(0.2), cursor: 'copy' },
-    onDropAccepted: upload
-  }, [({ openUploader }) => h(Fragment, [
-    div({ style: { flex: 'none', display: 'flex', alignItems: 'center', marginBottom: '1rem', justifyContent: 'flex-end' } }, [
-      h(Link, { onClick: download }, ['Download TSV']),
-      !Utils.editWorkspaceError(workspace) && h(Fragment, [
-        div({ style: { whiteSpace: 'pre' } }, ['  |  Drag or click to ']),
-        h(Link, { onClick: openUploader }, ['upload TSV'])
-      ]),
-      h(DelayedSearchInput, {
-        'aria-label': 'Search',
-        style: { width: 300, marginLeft: '1rem' },
-        placeholder: 'Search',
-        onChange: setTextFilter,
-        value: textFilter
-      })
-    ]),
-    div({ style: { flex: 1 } }, [
-      h(AutoSizer, [({ width, height }) => h(FlexTable, {
-        width, height, rowCount: amendedAttributes.length,
-        onScroll: y => saveScroll(0, y),
-        initialY,
-        hoverHighlight: true,
-        noContentMessage: _.isEmpty(initialAttributes) ? 'No Workspace Data defined' : 'No matching data',
-        columns: [{
-          size: { basis: 400, grow: 0 },
-          headerRenderer: () => h(HeaderCell, ['Key']),
-          cellRenderer: ({ rowIndex }) => editIndex === rowIndex ?
-            h(TextInput, {
-              'aria-label': 'Workspace data key',
-              autoFocus: true,
-              value: editKey,
-              onChange: setEditKey
-            }) :
-            renderDataCell(amendedAttributes[rowIndex][0], namespace)
-        }, {
-          size: { grow: 1 },
-          headerRenderer: () => h(HeaderCell, ['Value']),
-          cellRenderer: ({ rowIndex }) => {
-            const [originalKey, originalValue] = amendedAttributes[rowIndex]
-
-            return h(Fragment, [
-              div({ style: { flex: 1, minWidth: 0, display: 'flex' } }, [
-                editIndex === rowIndex ?
-                  h(TextInput, {
-                    'aria-label': 'Workspace data value',
-                    value: editValue,
-                    onChange: setEditValue
-                  }) :
-                  renderDataCell(originalValue, namespace)
-              ]),
-              editIndex === rowIndex ?
-                h(Fragment, [
-                  h(Select, {
-                    styles: { container: base => ({ ...base, marginLeft: '1rem', width: 150 }) },
-                    isSearchable: false,
-                    isClearable: false,
-                    menuPortalTarget: document.getElementById('root'),
-                    getOptionLabel: ({ value }) => _.startCase(value),
-                    value: editType,
-                    onChange: ({ value }) => setEditType(value),
-                    options: ['string', 'number', 'boolean', 'string list', 'number list', 'boolean list']
-                  }),
-                  h(Link, {
-                    tooltip: Utils.summarizeErrors(inputErrors) || 'Save changes',
-                    disabled: !!inputErrors.length,
-                    style: { marginLeft: '1rem' },
-                    onClick: () => saveAttribute(originalKey)
-                  }, [icon('success-standard', { size: 23 })]),
-                  h(Link, {
-                    tooltip: 'Cancel editing',
-                    style: { marginLeft: '1rem' },
-                    onClick: stopEditing
-                  }, [icon('times-circle', { size: 23 })])
-                ]) :
-                div({ className: 'hover-only' }, [
-                  h(Link, {
-                    disabled: !!Utils.editWorkspaceError(workspace),
-                    tooltip: Utils.editWorkspaceError(workspace) || 'Edit variable',
-                    style: { marginLeft: '1rem' },
-                    onClick: () => {
-                      setEditIndex(rowIndex)
-                      setEditValue(_.isObject(originalValue) ? originalValue.items.join(', ') : originalValue)
-                      setEditKey(originalKey)
-                      setEditType(_.isObject(originalValue) ? `${typeof originalValue.items[0]} list` : typeof originalValue)
-                    }
-                  }, [icon('edit', { size: 19 })]),
-                  h(Link, {
-                    'aria-label': 'Delete variable',
-                    disabled: !!Utils.editWorkspaceError(workspace),
-                    tooltip: Utils.editWorkspaceError(workspace) || 'Delete variable',
-                    style: { marginLeft: '1rem' },
-                    onClick: () => setDeleteIndex(rowIndex)
-                  }, [icon('trash', { size: 19 })])
-                ])
-            ])
-          }
-        }]
-      })])
-    ]),
-    !creatingNewVariable && editIndex === undefined && !Utils.editWorkspaceError(workspace) && h(FloatingActionButton, {
-      label: 'ADD VARIABLE',
-      iconShape: 'plus',
-      onClick: () => {
-        setEditIndex(initialAttributes.length)
-        setEditValue('')
-        setEditKey('')
-        setEditType('string')
-      }
-    }),
-    deleteIndex !== undefined && h(Modal, {
-      onDismiss: () => setDeleteIndex(),
-      title: 'Are you sure you wish to delete this variable?',
-      okButton: h(ButtonPrimary, {
-        onClick: _.flow(
-          withErrorReporting('Error deleting workspace variable'),
-          Utils.withBusyState(setBusy)
-        )(async () => {
-          setDeleteIndex()
-          await Ajax().Workspaces.workspace(namespace, name).deleteAttributes([amendedAttributes[deleteIndex][0]])
-          await loadAttributes()
-        })
-      },
-      'Delete Variable')
-    }, ['This will permanently delete the data from Workspace Data.']),
-    busy && spinnerOverlay
-  ])])
-}
 
 const ReferenceDataContent = ({ workspace: { workspace: { namespace, attributes } }, referenceKey, firstRender }) => {
   const [textFilter, setTextFilter] = useState('')
@@ -400,363 +123,9 @@ const ReferenceDataContent = ({ workspace: { workspace: { namespace, attributes 
   ])
 }
 
-const getDataset = dataExplorerUrl => {
-  // Either cohort was imported from standalone Data Explorer, eg
-  // https://test-data-explorer.appspot.com/
-  const dataset = _.find({ origin: new URL(dataExplorerUrl).origin }, datasets)
-  if (!dataset) {
-    // Or cohort was imported from embedded Data Explorer, eg
-    // https://app.terra.bio/#library/datasets/public/1000%20Genomes/data-explorer
-    const datasetName = unescape(dataExplorerUrl.split(/datasets\/(?:public\/)?([^/]+)\/data-explorer/)[1])
-    return _.find({ name: datasetName }, datasets)
-  }
-  return dataset
-}
-
-const ToolDrawer = _.flow(
-  Utils.withDisplayName('ToolDrawer'),
-  requesterPaysWrapper({
-    onDismiss: ({ onDismiss }) => onDismiss()
-  }),
-  withModalDrawer()
-)(({
-  workspace, workspace: { workspace: { bucketName, name: wsName, namespace, workspaceId } },
-  onDismiss, onIgvSuccess, onRequesterPaysError, entityMetadata, entityKey, selectedEntities
-}) => {
-  const [toolMode, setToolMode] = useState()
-  const [notebookNames, setNotebookNames] = useState()
-  const signal = Utils.useCancellation()
-
-  const { Buckets } = Ajax(signal)
-
-  Utils.useOnMount(() => {
-    const loadNotebookNames = _.flow(
-      withRequesterPaysHandler(onRequesterPaysError),
-      withErrorReporting('Error loading notebooks')
-    )(async () => {
-      const notebooks = await Buckets.listNotebooks(namespace, bucketName)
-      // slice removes 'notebooks/' and the .ipynb suffix
-      setNotebookNames(notebooks.map(notebook => notebook.name.slice(10, -6)))
-    })
-
-    loadNotebookNames()
-  })
-
-  const entitiesCount = _.size(selectedEntities)
-  const isCohort = entityKey === 'cohort'
-
-  const dataExplorerButtonEnabled = isCohort && entitiesCount === 1 && _.values(selectedEntities)[0].attributes.data_explorer_url !== undefined
-  const origDataExplorerUrl = dataExplorerButtonEnabled ? _.values(selectedEntities)[0].attributes.data_explorer_url : undefined
-  const [baseURL, urlSearch] = origDataExplorerUrl ? origDataExplorerUrl.split('?') : []
-  const dataExplorerUrl = origDataExplorerUrl && `${baseURL}?${qs.stringify({ ...qs.parse(urlSearch), wid: workspaceId })}`
-  const openDataExplorerInSameTab = dataExplorerUrl && (dataExplorerUrl.includes('terra.bio') || _.some({ origin: new URL(dataExplorerUrl).origin }, datasets))
-  const dataset = openDataExplorerInSameTab && getDataset(dataExplorerUrl)
-  const linkBase = openDataExplorerInSameTab && Nav.getLink(dataset.authDomain ? 'data-explorer-private' : 'data-explorer-public', { dataset: dataset.name })
-  const dataExplorerPath = openDataExplorerInSameTab && `${linkBase}?${dataExplorerUrl.split('?')[1]}`
-
-  const notebookButtonEnabled = isCohort && entitiesCount === 1
-
-  const { title, drawerContent } = Utils.switchCase(toolMode, [
-    'IGV', () => ({
-      title: 'IGV',
-      drawerContent: h(IGVFileSelector, {
-        onSuccess: onIgvSuccess,
-        selectedEntities
-      })
-    })
-  ], [
-    'Workflow', () => ({
-      title: 'YOUR WORKFLOWS',
-      drawerContent: h(WorkflowSelector, { workspace, selectedEntities })
-    })
-  ], [
-    'Notebook', () => ({
-      drawerContent: h(NotebookCreator, {
-        bucketName, namespace,
-        existingNames: notebookNames,
-        onSuccess: async (notebookName, notebookKernel) => {
-          const cohortName = _.values(selectedEntities)[0].name
-          const contents = notebookKernel === 'r' ? cohortRNotebook(cohortName) : cohortNotebook(cohortName)
-          await Buckets.notebook(namespace, bucketName, notebookName).create(JSON.parse(contents))
-          Nav.goToPath('workspace-notebook-launch', { namespace, name: wsName, notebookName: `${notebookName}.ipynb` })
-        },
-        onDismiss: () => setToolMode(undefined),
-        reloadList: _.noop
-      })
-    })
-  ], [
-    Utils.DEFAULT, () => ({
-      title: 'OPEN WITH...',
-      drawerContent: h(Fragment, [
-        div({ style: Style.modalDrawer.content }, [
-          div([
-            h(ModalToolButton, {
-              onClick: () => setToolMode('IGV'),
-              disabled: isCohort,
-              tooltip: isCohort ? 'IGV cannot be opened with cohorts' : 'Open with Integrative Genomics Viewer',
-              icon: igvLogo,
-              text: 'IGV'
-            }),
-            h(ModalToolButton, {
-              onClick: () => setToolMode('Workflow'),
-              disabled: isCohort,
-              tooltip: isCohort ? 'Workflow cannot be opened with cohorts' : 'Open with Workflow',
-              icon: wdlLogo,
-              text: 'Workflow'
-            }),
-            h(ModalToolButton, {
-              onClick: !openDataExplorerInSameTab ? onDismiss : undefined,
-              href: openDataExplorerInSameTab ? dataExplorerPath : dataExplorerUrl,
-              ...(!openDataExplorerInSameTab ? Utils.newTabLinkProps : {}),
-              disabled: !dataExplorerButtonEnabled,
-              tooltip: Utils.cond(
-                [!entityMetadata.cohort, () => 'Talk to your dataset owner about setting up a Data Explorer. See the "Making custom cohorts with Data Explorer" help article.'],
-                [isCohort && entitiesCount > 1, () => 'Select exactly one cohort to open in Data Explorer'],
-                [isCohort && !dataExplorerUrl, () => 'Cohort is too old, please recreate in Data Explorer and save to Terra again'],
-                [!isCohort, () => 'Only cohorts can be opened with Data Explorer']
-              ),
-              icon: dataExplorerLogo,
-              text: 'Data Explorer'
-            }),
-            h(ModalToolButton, {
-              onClick: () => setToolMode('Notebook'),
-              disabled: !notebookButtonEnabled,
-              tooltip: Utils.cond(
-                [!entityMetadata.cohort, () => 'Unable to open with notebooks. See the "Making custom cohorts with Data Explorer" help article for more details.'],
-                [isCohort && entitiesCount > 1, () => 'Select exactly one cohort to open in notebook'],
-                [!isCohort, () => 'Only cohorts can be opened with notebooks'],
-                [notebookButtonEnabled, () => 'Create a Python 2 or 3 notebook with this cohort']
-              ),
-              icon: jupyterLogo,
-              text: 'Notebook'
-            })
-          ])
-        ])
-      ])
-    })
-  ])
-
-  return div({ style: { padding: '1.5rem', display: 'flex', flexDirection: 'column', flex: 1 } }, [
-    h(TitleBar, {
-      title,
-      onPrevious: toolMode ? () => { setToolMode(undefined) } : undefined,
-      onDismiss
-    }),
-    div({
-      style: {
-        borderRadius: '1rem',
-        border: `1px solid ${colors.dark(0.5)}`,
-        padding: '0.25rem 0.875rem',
-        margin: '0.5rem 0 2rem',
-        alignSelf: 'flex-start',
-        fontSize: 12
-      }
-    }, [
-      `${entitiesCount} ${entityKey + (entitiesCount > 1 ? 's' : '')} selected`
-    ]),
-    drawerContent
-  ])
-})
-
 const SnapshotContent = ({ snapshotDetails, snapshotKey: [snapshotName] }) => {
   // ToDo: Add rendering of snapshot tables
   return h(SnapshotInfo, { snapshotId: snapshotDetails[snapshotName].id, snapshotName })
-}
-
-const EntitiesContent = ({
-  workspace, workspace: { workspace: { namespace, name, attributes: { 'workspace-column-defaults': columnDefaults } }, workspaceSubmissionStats: { runningSubmissionsCount } },
-  entityKey, entityMetadata, loadMetadata, firstRender
-}) => {
-  // State
-  const [selectedEntities, setSelectedEntities] = useState({})
-  const [deletingEntities, setDeletingEntities] = useState(false)
-  const [copyingEntities, setCopyingEntities] = useState(false)
-  const [nowCopying, setNowCopying] = useState(false)
-  const [refreshKey, setRefreshKey] = useState(0)
-  const [showToolSelector, setShowToolSelector] = useState(false)
-  const [igvFiles, setIgvFiles] = useState(undefined)
-  const [igvRefGenome, setIgvRefGenome] = useState('')
-
-  const downloadForm = useRef()
-
-  // Render helpers
-  const renderDownloadButton = columnSettings => {
-    const disabled = entityKey.endsWith('_set_set')
-    return h(Fragment, [
-      form({
-        ref: downloadForm,
-        action: `${getConfig().orchestrationUrlRoot}/cookie-authed/workspaces/${namespace}/${name}/entities/${entityKey}/tsv`,
-        method: 'POST'
-      }, [
-        input({ type: 'hidden', name: 'FCtoken', value: getUser().token }),
-        input({ type: 'hidden', name: 'attributeNames', value: _.map('name', _.filter('visible', columnSettings)).join(',') }),
-        input({ type: 'hidden', name: 'model', value: 'flexible' })
-      ]),
-      h(ButtonPrimary, {
-        disabled,
-        tooltip: disabled ?
-          'Downloading sets of sets as TSV is not supported at this time' :
-          `Download a .tsv file containing all the ${entityKey}s in this table`,
-        onClick: () => downloadForm.current.submit()
-      }, [
-        icon('download', { style: { marginRight: '0.5rem' } }),
-        'Download all Rows'
-      ])
-    ])
-  }
-
-  const buildTSV = (columnSettings, entities) => {
-    const sortedEntities = _.sortBy('name', entities)
-    const isSet = _.endsWith('_set', entityKey)
-    const setRoot = entityKey.slice(0, -4)
-    const attributeNames = _.flow(
-      _.filter('visible'),
-      _.map('name'),
-      isSet ? _.without([`${setRoot}s`]) : _.identity
-    )(columnSettings)
-
-    const entityTsv = Utils.makeTSV([
-      [`entity:${entityKey}_id`, ...attributeNames],
-      ..._.map(({ name, attributes }) => {
-        return [name, ..._.map(attribute => Utils.entityAttributeText(attributes[attribute], true), attributeNames)]
-      }, sortedEntities)
-    ])
-
-    if (isSet) {
-      const membershipTsv = Utils.makeTSV([
-        [`membership:${entityKey}_id`, setRoot],
-        ..._.flatMap(({ attributes, name }) => {
-          return _.map(({ entityName }) => [name, entityName], attributes[`${setRoot}s`].items)
-        }, sortedEntities)
-      ])
-
-      const zipFile = new JSZip()
-        .file(`${entityKey}_entity.tsv`, entityTsv)
-        .file(`${entityKey}_membership.tsv`, membershipTsv)
-
-      return zipFile.generateAsync({ type: 'blob' })
-    } else {
-      return entityTsv
-    }
-  }
-
-  const renderCopyButton = (entities, columnSettings) => {
-    return h(Fragment, [
-      h(ButtonPrimary, {
-        style: { marginLeft: '1rem' },
-        tooltip: `Copy only the ${entityKey}s visible on the current page to the clipboard in .tsv format`,
-        onClick: _.flow(
-          withErrorReporting('Error copying to clipboard'),
-          Utils.withBusyState(setNowCopying)
-        )(async () => {
-          const str = buildTSV(columnSettings, entities)
-          await clipboard.writeText(str)
-          notify('success', 'Successfully copied to clipboard', { timeout: 3000 })
-        })
-      }, [
-        icon('copy-to-clipboard', { style: { marginRight: '0.5rem' } }),
-        'Copy Page to Clipboard'
-      ]),
-      nowCopying && spinner()
-    ])
-  }
-
-  const renderSelectedRowsMenu = columnSettings => {
-    const isSet = _.endsWith('_set', entityKey)
-    const noEdit = Utils.editWorkspaceError(workspace)
-    const disabled = entityKey.endsWith('_set_set')
-
-    return !_.isEmpty(selectedEntities) && h(PopupTrigger, {
-      side: 'bottom',
-      closeOnClick: true,
-      content: h(Fragment, [
-        h(MenuButton, {
-          disabled,
-          tooltip: disabled ?
-            'Downloading sets of sets as TSV is not supported at this time' :
-            `Download the selected data as a file`,
-          onClick: async () => {
-            const tsv = buildTSV(columnSettings, selectedEntities)
-            isSet ?
-              FileSaver.saveAs(await tsv, `${entityKey}.zip`) :
-              FileSaver.saveAs(new Blob([tsv], { type: 'text/tab-separated-values' }), `${entityKey}.tsv`)
-          }
-        }, ['Download as TSV']),
-        h(MenuButton, {
-          tooltip: 'Open the selected data to work with it',
-          onClick: () => setShowToolSelector(true)
-        }, ['Open with...']),
-        h(MenuButton, {
-          tooltip: 'Send the selected data to another workspace',
-          onClick: () => setCopyingEntities(true)
-        }, ['Export to Workspace']),
-        h(MenuButton, {
-          tooltip: noEdit ? 'You don\'t have permission to modify this workspace' : 'Permanently delete the selected data',
-          disabled: noEdit,
-          onClick: () => setDeletingEntities(true)
-        }, ['Delete Data'])
-      ])
-    }, [h(Link, { style: { marginRight: '1rem' } }, [
-      icon('ellipsis-v-circle', { size: 24 })
-    ])])
-  }
-
-  // Render
-  const { initialX, initialY } = firstRender ? StateHistory.get() : {}
-  const selectedKeys = _.keys(selectedEntities)
-  const selectedLength = selectedKeys.length
-
-  return igvFiles ?
-    h(IGVBrowser, { selectedFiles: igvFiles, refGenome: igvRefGenome, workspace, onDismiss: () => setIgvFiles(undefined) }) :
-    h(Fragment, [
-      h(DataTable, {
-        persist: true, firstRender, refreshKey, editable: !Utils.editWorkspaceError(workspace),
-        entityType: entityKey, entityMetadata, columnDefaults, workspaceId: { namespace, name },
-        onScroll: saveScroll, initialX, initialY,
-        selectionModel: {
-          selected: selectedEntities,
-          setSelected: setSelectedEntities
-        },
-        childrenBefore: ({ entities, columnSettings }) => div({
-          style: { display: 'flex', alignItems: 'center', flex: 'none' }
-        }, [
-          renderDownloadButton(columnSettings),
-          !_.endsWith('_set', entityKey) && renderCopyButton(entities, columnSettings),
-          div({ style: { margin: '0 1.5rem', height: '100%', borderLeft: Style.standardLine } }),
-          div({ style: { marginRight: '0.5rem' } }, [`${selectedLength} row${selectedLength === 1 ? '' : 's'} selected`]),
-          renderSelectedRowsMenu(columnSettings)
-        ])
-      }),
-      deletingEntities && h(EntityDeleter, {
-        onDismiss: () => setDeletingEntities(false),
-        onSuccess: () => {
-          setDeletingEntities(false)
-          setSelectedEntities({})
-          setRefreshKey(_.add(1))
-          loadMetadata()
-        },
-        namespace, name,
-        selectedEntities: selectedKeys, selectedDataType: entityKey, runningSubmissionsCount
-      }),
-      copyingEntities && h(ExportDataModal, {
-        onDismiss: () => setCopyingEntities(false),
-        workspace,
-        selectedEntities: selectedKeys, selectedDataType: entityKey, runningSubmissionsCount
-      }),
-      h(ToolDrawer, {
-        workspace,
-        isOpen: showToolSelector,
-        onDismiss: () => setShowToolSelector(false),
-        onIgvSuccess: ({ selectedFiles, refGenome }) => {
-          setShowToolSelector(false)
-          setIgvFiles(selectedFiles)
-          setIgvRefGenome(refGenome)
-        },
-        entityMetadata,
-        entityKey,
-        selectedEntities
-      })
-    ])
 }
 
 const DeleteObjectModal = ({ name, workspace: { workspace: { namespace, bucketName } }, onSuccess, onDismiss }) => {
