+++ src/pages/workspaces/workspace/workflows/DataStepContent.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { Component, Fragment } from 'react'
 import { div, h } from 'react-hyperscript-helpers'
 import { ButtonPrimary, IdContainer, RadioButton } from 'src/components/common'
-import DataTable from 'src/components/DataTable'
+import DataTable from 'src/components/data/DataTable'
 import { ValidatedInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
 import { FormLabel } from 'src/libs/forms'
